Revert default AudioOptions change for webrtc_voice_engine.cc on Android (#205)
These default values got changed from true to false in #178. On Android,
the way the audio options are set is fairly janky, and actually relied
on these initially defaulting to true.

I'm not sure if other non iOS platforms rely on these initial default
values as well, but that's something to be aware of for those platforms.